refactor: split settings tabs into components
